+++ pages/display_products.php
@@ -1,20 +1,26 @@
 <?php 
 include "header.php";
 include "sidenav.php";
+if(!isset($_SESSION['user']) || count($_SESSION['user'])<=0){
+		redirect("../index.php");
+}
 $query = "select * from product";
+mysql_set_charset('utf8',$conn);
 	$result = mysql_query($query,$conn);
+
+
 	if(isset($_REQUEST['msg'])){
 			if($_REQUEST['msg']=="success"){
 ?>
 				<div class="alert alert-danger">
-                    <strong><?php echo $msg = "Product is deleted successfully!"; ?></strong>
-                </div>
+					<strong><?php echo $msg = "Product is deleted successfully!"; ?></strong>
+				</div>
 <?php
 		}else if($_REQUEST['msg']=="failure"){
 ?>
 				<div class="alert alert-warning">
-                    <strong><?php echo $msg = "Product is not deleted!"; ?></strong>
-                </div>
+					<strong><?php echo $msg = "Product is not deleted!"; ?></strong>
+				</div>
 <?php
 		}
 	}else{
@@ -23,50 +29,57 @@ $query = "select * from product";
 ?>
  <div id="page-wrapper">
 
-            <div class="container-fluid">
+			<div class="container-fluid">
 
-                <!-- Page Heading -->
-        <div class="row">
-                    <div class="col-lg-12">
-                        <h1 class="page-header">
-                            <img src="../vraj1.png" title="logo" alt="logo" height="70">
-                            <small>Display Products</small>
-                        </h1>
-                        <ol class="breadcrumb">
-                            <li>
-                                <i class="fa fa-home"></i>  <a href="home.php">Dashboard</a>
-                            </li>
-                            <li class="active">
+				<!-- Page Heading -->
+		<div class="row">
+					<div class="col-lg-12">
+						<h1 class="page-header">
+							<img src="../vraj1.png" title="logo" alt="logo" height="70">
+							<small>Display Products</small>
+						</h1>
+						<ol class="breadcrumb">
+							<li>
+								<i class="fa fa-home"></i>  <a href="home.php">Dashboard</a>
+							</li>
+							<li class="active">
 							Manage Products
-                            </li>
+							</li>
 							<li class="active">
 							<i class="fa fa-fw fa-table"></i>
 							Display Products
-                            </li>
-                        </ol>
-                    </div>
-                </div>
-                <!-- /.row -->
-	 		<div class="row">
-                    <div class="col-lg-12">            
-                        <div class="table-responsive"><!--table  div class-->
+							</li>
+						</ol>
+					</div>
+				</div>
+				<!-- /.row -->
+			<div class="row">
+					<div class="col-lg-12">            
+						<div class="table-responsive"><!--table  div class-->
 							<table class="table table-hover" ><!--table class-->
 							<tr>
 								<th>Photo</th>
 								<th>Name</th>
 								<th>Description</th>
-								<th>Category name</th>
-								<th>Price</th>
-								<th>Edit</th>
-								<th>Delete</th>
+								<th nowrap>Category name</th>
+								<th nowrap>Price</th>
+								<th nowrap>Status</th>
+								<th nowrap><i class="fa fa-fw fa-edit"></i> Edit</th>
+								<th nowrap><i class="fa fa-trash-o"></i> Delete</th>
 							</tr>
 							<?php
 							while($row = mysql_fetch_array($result))
 							{
+								if($row['product_status']==1){
+								$status='<font color="limegreen"><b> ACTIVE</font></b>';
+							}
+							else{
+								$status='<font color="crimson"><b>INACTIVE</font></b>';
+							}
 							?>
 							<tr>
-								<td><img class="img-thumbnail" src="<?php if(!$row['product_image']){
-									echo "http://placehold.it/400x400";} else{ echo $row['product_image'];
+								<td><img class="img-thumbnail" width="100" height="100" src="<?php if(!$row['product_image']){
+									echo "http://placehold.it/200x200";} else{ echo $row['product_image'];
 								}?>"alt="">
 								</td>
 								<td><?php echo strtoupper($row['product_name']); ?></td>
@@ -78,13 +91,13 @@ $query = "select * from product";
 							$cat_result = mysql_query($cat_query,$conn);
 							$cat_row = mysql_fetch_array($cat_result)
 						?>
-								<td><?php echo $cat_row['name'] ?></td>
+								<td nowrap><?php echo $cat_row['name'] ?></td>
 
 
-								<td><?php echo $row['price']; ?>INR</td>
-								
-								<td><a href="edit_product.php?pid=<?php echo $row['product_id']; ?>">Edit</a></td>
-								<td><a href="delete_product.php?pid=<?php echo $row['product_id']; ?>">Delete</a></td>
+								<td nowrap><?php echo $row['price']; ?><i class="fa fa-rupee"></i></td>
+								<td nowrap><?php echo $status; ?></td>
+								<td nowrap><a href="edit_product.php?pid=<?php echo $row['product_id']; ?>"><i class="fa fa-fw fa-edit"></i> Edit</a></td>
+								<td nowrap><a onClick="javascript: return confirm('Are you sure you want to delete the product?');" href="delete_product.php?pid=<?php echo $row['product_id']; ?>"><i class="fa fa-trash-o"></i> Delete</a></td>
 							</tr>
 							<?php
 							}
@@ -93,11 +106,11 @@ $query = "select * from product";
 
 						</div>
 </center>
-            </div>
-            <!-- /.container-fluid -->
+			</div>
+			<!-- /.container-fluid -->
 
-        </div>
-        <!-- /#page-wrapper -->
+		</div>
+		<!-- /#page-wrapper -->
 <?php 
 include "footer.php";
 ?>
